[FEAT/#99] id 데이터 형식 변경

+++ data/src/main/java/com/terning/data/datasource/ScrapDataSource.kt
@@ -3,5 +3,8 @@ package com.terning.data.datasource
 import com.terning.data.dto.NonDataBaseResponse
 
 interface ScrapDataSource {
-    suspend fun postScrap(id: Int, color: Int): NonDataBaseResponse
+    suspend fun postScrap(
+        id: Long,
+        color: Int,
+    ): NonDataBaseResponse
 }
