+++ domain/src/main/java/com/terning/domain/repository/ScrapRepository.kt
@@ -1,5 +1,5 @@
 package com.terning.domain.repository
 
 interface ScrapRepository {
-    suspend fun postScrap(id: Int, color: Int)
+    suspend fun postScrap(id: Long, color: Int)
 }
